Disable use of JSC signal based options when mach exception handler use is not allowed. https://bugs.webkit.org/show_bug.cgi?id=278401 rdar://134053581
Reviewed by Keith Miller and Yusuke Suzuki.

If the sandbox policy is known for mach exception handler use, we'll either allow or block it
in accordance with the sandbox policy.

if the sandbox policy is not known, we will use JIT disablement as a proxy indicator that use
of mach exception handler use is not allowed in the system.  So, in thos case, we'll disable
JSC signal based options when the JIT is disabled.

Note that all places where we disable the JIT (via setting useJIT to false) will be followed by
a call to Options::notifyOptionsChanged().  So, that is the natural place to disable signal
handler based options for this purpose if appropriate.

* Source/JavaScriptCore/API/tests/testapi.cpp:
(configureJSCForTesting):
* Source/JavaScriptCore/jsc.cpp:
(jscmain):
* Source/JavaScriptCore/runtime/Options.cpp:
(JSC::disableAllSignalHandlerBasedOptions):
(JSC::Options::notifyOptionsChanged):
* Source/JavaScriptCore/runtime/Options.h:
* Source/WebKit/Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceEntryPoint.h:
(WebKit::XPCServiceInitializer):
* Source/WebKit/Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceEntryPoint.mm:
(WebKit::setJSCOptions):
* Source/WebKit/WebProcess/EntryPoint/Cocoa/XPCService/WebContentServiceEntryPoint.mm:
(WEBCONTENT_SERVICE_INITIALIZER):
* Tools/DumpRenderTree/mac/DumpRenderTree.mm:
(dumpRenderTree):

Canonical link: https://commits.webkit.org/282592@main