[Grid] Start considering grid item logical min width when computing inline size contribution. https://bugs.webkit.org/show_bug.cgi?id=260287 rdar://113984636
Reviewed by Alan Baradlay.

GridTrackSizingAlgorithm::minContentForGridItem is supposed to compute
the min content contribution for the grid item which is supposed to
take into consideration the grid item's minimum size property in the
appropriate axis.

It currently does not do that so this patch is a first step towards that
by taking into consideration the minimum logical width if it is either
fixed or max-content.

The grid-intrinsic-maximums test was testing the min-width: max-content
case specifically and had 2 different versions: one from WPT and one
in fast/css-grid-layout. We were passing the latter but the test
expectations were incorrect while the WPT was ok. So I deleted the one
in fast/css-grid-layout since they would otherwise be the same.

* LayoutTests/TestExpectations:
* LayoutTests/fast/css-grid-layout/grid-intrinsic-maximums-expected.html: Removed.
* LayoutTests/fast/css-grid-layout/grid-intrinsic-maximums.html: Removed.
* Source/WebCore/rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::GridTrackSizingAlgorithmStrategy::minContentForGridItem const):

Canonical link: https://commits.webkit.org/282616@main