[JSC] Should preserve the index in compileHasIndexedProperty if needed https://bugs.webkit.org/show_bug.cgi?id=278427 rdar://129328633
Reviewed by Keith Miller.

The EnumeratorNextUpdateIndexAndMode node is expected to return two results.
In the fast path for case IndexedMode, the index value should be returned as
one of the results. However, the slow path of compileHasIndexedProperty
triggers a operation call with a single result. So, compileHasIndexedProperty
should preserve the index in compileHasIndexedProperty if needed.

* Source/JavaScriptCore/dfg/DFGSpeculativeJIT.cpp:
* Source/JavaScriptCore/dfg/DFGSpeculativeJIT.h:

Canonical link: https://commits.webkit.org/280938.256@safari-7619-branch