[Sonoma wk2] fast/replaced/pdf-as-embed-with-no-mime-type-is-not-blank.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=278429 rdar://134377511
Reviewed by Wenson Hsieh.

PDFKit performs an animated fade-in of the PDF document to render, and
as such it looks like the <embed> just never loads. We work around this
issue by rAF-ing after a timeout of 250ms, which we've previously
determined to be enough time for the PDF fade-in animation to pass.

Instead of adjusting the single failing test, though, we introduce a new
utility UIHelper.waitForPDFFadeIn, and adopt that in other relevant
tests.

Also, we get rid of the PluginsEnabled test preference, since we removed
support for that key in 274917@main.

* LayoutTests/compositing/plugins/pdf/pdf-in-embed-rounded-border.html:
* LayoutTests/compositing/plugins/pdf/pdf-in-embed-with-root-page-zoom.html:
* LayoutTests/compositing/plugins/pdf/pdf-in-embed.html:
* LayoutTests/compositing/plugins/pdf/pdf-in-styled-embed.html:
* LayoutTests/compositing/plugins/pdf/pdf-scrolling-tree.html:
* LayoutTests/fast/replaced/pdf-as-embed-with-no-mime-type-is-not-blank.html:
* LayoutTests/http/tests/pdf/linearized-pdf-in-iframe.html:
* LayoutTests/platform/mac-wk2/TestExpectations:
* LayoutTests/resources/ui-helper.js:
(window.UIHelper.async waitForPDFFadeIn):
(window.UIHelper):

Canonical link: https://commits.webkit.org/282638@main