RELEASE_ASSERT that ObjectIdentifier is valid when constructed from a raw value https://bugs.webkit.org/show_bug.cgi?id=278522
Reviewed by Darin Adler.

We recently made it so that ObjectIdentifier no longer has a default constructor
or an "invalid" state. However, a call site could still in theory construct an
ObjectIdentifier with a raw value of 0 to construct an invalid identifier. Doing
so would only hit an assertion in debug builds. For hardening, I'm suggesting we
turn this debug assertion into a release one. This tested as performance neutral.

* Source/WTF/wtf/ObjectIdentifier.h:
(WTF::ObjectIdentifierGeneric::generate):
(WTF::ObjectIdentifierGeneric::ObjectIdentifierGeneric):

Canonical link: https://commits.webkit.org/282646@main