[WebGPU] Add additional validation to memcpySpan call in Queue::writeBuffer https://bugs.webkit.org/show_bug.cgi?id=278448 rdar://132230239
Reviewed by Dan Glastonbury.

We were creating a span without respect to the buffer length because
it was already guarnateed in the validateWriteBuffer that buffer.length
was greater than data.size() + offset.

However a bug or change to the validation function could result in that
logic being skipped.

Ensure the span is within the range of the buffer by creating another
span from the buffer conents + length.

* Source/WebGPU/WebGPU/Queue.mm:
(WebGPU::Queue::writeBuffer):

Canonical link: https://commits.webkit.org/282675@main