Fix assertion after 281812@main https://bugs.webkit.org/show_bug.cgi?id=278682 rdar://134731526
Reviewed by Timothy Hatcher.

My assertion had an off-by-one error that was usually true because
it was checking that uninitialized memory was 0.  JSStringGetUTF8CString
returns a length that includes the null termination character, which
I got correct in the call to WKStringCreateWithUTF8CStringWithLength
but not in the assertion.  This fixes the assertion and makes the
tests stop asserting in debug builds.

* Source/WebKit/Shared/API/APISerializedScriptValue.cpp:
(API::valueToWKObject):

Canonical link: https://commits.webkit.org/282769@main