Fix assert related to preference update https://bugs.webkit.org/show_bug.cgi?id=278689 rdar://134746970
Reviewed by Chris Dumez and Richard Robinson.

In some cases we hit an assert when updating preferences sent from the UI process. This can happen
when the preference has been set by an underlying framework in the volatile preference domain. The
value we get when reading the preference will not be the expected one, then, since the volatile
domain comes very early in the preference search order. This patch addresses this by adding an
exception for this preference key.

* Source/WebKit/Shared/Cocoa/AuxiliaryProcessCocoa.mm:
(WebKit::AuxiliaryProcess::setPreferenceValue):

Canonical link: https://commits.webkit.org/282773@main