xhr/resources/echo-headers.py is triggering flaky load failures rdar://92818673 https://bugs.webkit.org/show_bug.cgi?id=278708
Reviewed by Anne van Kesteren.

Without the Content-Length header, Safari (and sometimes Chrome, not Firefox) are failing the load when TCP connection gets closed.
Fix the server script to workaround this issue as the test is not about this particular issue.
We will in parallel further investigate what should be done with this case.

* LayoutTests/imported/w3c/web-platform-tests/xhr/resources/echo-headers.py:
(main):

Canonical link: https://commits.webkit.org/282784@main