REGRESSION(281966@main): [GTK] can't drag and drop files to a gitlab issue https://bugs.webkit.org/show_bug.cgi?id=278644
Reviewed by Adrian Perez de Castro.

The problem is that WebPageProxy::performDragOperation() no longer sends
the request to the web process immediately, but after sending a message
to the network process. In GTK the leave notification is emitted right
after the drop, so we end up sending the exit operation to the web
process before the actual drop operation. We don't really need to wait
until the paths have been allowed in the network process to continue
with the drag operation. Once the operation is performed and the upload
request is sent to the network process, the previous message to allow
the file has already been handled. Also make sure we only grant access
to files when the operation is going to be performed, not for enter,
update or exit operations.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::performDragOperation):
(WebKit::WebPageProxy::performDragControllerAction):

Canonical link: https://commits.webkit.org/282837@main