Take a background assertion for processes having muted capture rdar://130691389 https://bugs.webkit.org/show_bug.cgi?id=278560
Reviewed by Chris Dumez.

When a process has muted capture, it may get quickly suspended.
This for instance happens when switching to another tab in Safari.
This also happens for any WKWebView application that will go to the background, even though the WKWebView application may not get suspended.

We now take a background assertion for any web process that has muted capture.
This will ensure that the web process does not get suspended, as long as the UIProcess is not suspended.
This assertion will not delay UIProcess suspension, contrary to the past incarnation of this patch.

Manually tested by observing that a WebRTC connection continues sending black frames in a backgrounded tab on iOS.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::updateThrottleState):
* Source/WebKit/UIProcess/WebPageProxy.h:

Canonical link: https://commits.webkit.org/282845@main