[ews-build] Highlight flaky API test failures when tests fails in first run but passes in rerun https://bugs.webkit.org/show_bug.cgi?id=278742
Reviewed by Ryan Haddad.

When the first run has test failures, but second run for exact same code passes all
the tests, it means the test failures in first run were flaky. Indicate that in buildbot
build summary, so that they are easier to notice and fix. We already do that for layout-tests.

* Tools/CISupport/ews-build/steps.py:
(RunAPITests.run):

Canonical link: https://commits.webkit.org/282861@main